Atricle rate methods, add logs in services

+++ IdentityAut/DTOs/DTOs/Article/ArticleDTO.cs
@@ -11,7 +11,7 @@ namespace Core.DTOs.Article
         public Int32 Id { get; set; }
         public String Title { get; set; }
         public String ArticlePicture { get; set; }
-        public Int32 PositiveRate { get; set; }
+        public Double PositiveRate { get; set; }
         public DateTime DateTime { get; set; }
         public String ShortDescription { get; set; }
         public String ArticleSourceUrl { get; set; }
